refactor(git-changes): extract openGitChanges out of workspace.store
Task 2 added openGitChanges to workspace.store.ts, pushing it from 496 to 514
lines and over the 500-line hard cap (god-file check failed in pnpm check).
Move the singleton-open logic to a dedicated panel-kits/git-changes/open-git-changes.ts
function (more cohesive — the logic belongs to the git-changes domain), and have
the command handler call it. Store is back to 494 lines. Regression was missed
because Task 2 ran typecheck but not the full pnpm check.
